[TIMOB-23903] Build against V8 5.7.492.71 (#8382)
* Use V8 5.3.332.41

* Move from WeakCallbackData to WeakCallbackInfo API, since the former is removed now.

* Update calls to Template#Set(). Can only set primitives or other Templates, so the Set*Method calls needed to set the generated FunctionTemplate (and not it's function). Also need to hack around hanging an External wrapping a jclass on the __javaClass_ property by using a natiev data property getter callback that gets the External passed as Data and returning that.

* Bump module api version for Andorid to 4

* Rebuild the ti.map and ti.facebook native modules for Android with higher version/moduleApiVersion against a build of the SDK

* Add scons binary script

* Getting closer to working, mocha tests crash in the middle not at startup now.

* Hmm, now we crash with a reference table overflow in JNI for PlatformModule...

* Closer still. Crashes on Ti.UI.Label.apiname test, because getproxyTemplate must be throwing some uncaught excpetion, so the toLocalChecked() on GetFunction is failing.

* Fix double listing of property

* Add exception handling for cases where we blindly called ToLocalChecked() before

* Try to set up so when JVM collects a weak reference to a Java object with a proxy that we then go back and kill the proxy/JS object on the C++ side. Tehre's still an issue here where we're not cleaning up jclass references somehow that I need to fix...

* Booya! Got it to stop crashing on the nasty Ti.UI.Window test

* Fix duplicate property name errors on WebView proxy

* Fix local ref leak, guard against trying to reset the Persistent twice in destructor

* Be more paranoid about cleaning up local refs in Proxy for jobjects

* Clean up docs for ReferenceTable

* Start rolling back some debug stuff

* Remove FIXME that no longer applies, lower log level of debug log

* Remove FIXME that no longer applies

* lower log level of debug log

* Add note about possibly improving raw pointer usage across the JNI boundary to Proxy instances

* Re-enable most of the XML tests now that we don't have JNI leaks so they don't *all* barf/crash

* Roll back the reference queue/proxy cleanup stuff on ReferenceTable. General cleanup

* use longs for keys in ReferenceTable, not ints. Clean up header includes.

* USe non-deprecated version of v8::Debug::ProcessDebugMessages()

* Update to V8 5.7.492.71

* Fix reference to V8Util to be titanium::V8Util

+++ android/runtime/v8/src/native/ReferenceTable.h
@@ -1,6 +1,6 @@
 /**
  * Appcelerator Titanium Mobile
- * Copyright (c) 2012 by Appcelerator, Inc. All Rights Reserved.
+ * Copyright (c) 2016-2017 by Appcelerator, Inc. All Rights Reserved.
  * Licensed under the terms of the Apache Public License
  * Please see the LICENSE included with this distribution for details.
  */
@@ -24,11 +24,11 @@ namespace titanium {
 class ReferenceTable
 {
 public:
-	static jint createReference(jobject object);
-	static void destroyReference(jint key);
-	static void makeWeakReference(jint key);
-	static jobject clearWeakReference(jint key);
-	static jobject getReference(jint key);
+	static jlong createReference(jobject object);
+	static void destroyReference(jlong key);
+	static void makeWeakReference(jlong key);
+	static jobject clearWeakReference(jlong key);
+	static jobject getReference(jlong key);
 };
 
 } // namespace titanium
